+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - The FAB plugin version no longer opts out of Crashpad handler configuration ([#1028](https://github.com/getsentry/sentry-unreal/pull/1028))
+- The FAB version of the plugin builds successfully on Android and iOS ([#1027](https://github.com/getsentry/sentry-unreal/pull/1027))
 
 ## 1.0.0-beta.6